backend/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
env==0.1.0
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2020.4
sqlparse==0.4.1
